refactor(server): migrate server adapters to class-via-configurable; delete configurableFn
RALPH: PRD #27, issue #45

Lift the `Server` interface to an abstract class and rewrite
`ExpressServer` and `FastifyServer` as
`class extends configurable(serverConfigPipe, Server)`.
Delete the deprecated `configurableFn` from `src/utilities/configurable.ts`.

Key decisions:
- `Server` abstract class folds in the shared logic from the `base()`
  factory (route registration queue, schema validation, socket/cors
  setup, error handling, OpenAPI integration) as concrete methods
- Framework-specific behaviour (request parsing, response serialisation,
  route/error/404 registration, server start) are protected abstract
  methods overridden by each leaf class
- `Server` uses a protected `setup(httpServer, config)` two-phase init
  called from leaf constructors (after `super(config)` so `this.config`
  is available), avoiding the need for base-class constructor args that
  would leak into `create()`'s public signature
- Both leaf classes use `protected constructor` with `static create`
  via inherited `configurable` — no nested anonymous classes
- `start()` parameter dropped (was declared but unused; `config.port`
  is the authoritative source)

Files changed:
- src/server/adapters/base.ts (interface → abstract class with base()
  logic)
- src/server/adapters/express/index.ts (configurableFn → class)
- src/server/adapters/fastify/index.ts (configurableFn → class)
- src/server/adapters/utils.ts (deleted — logic moved to Server class)
- src/utilities/configurable.ts (configurableFn removed)
