+++ src/server/adapters/fastify/index.ts
@@ -11,113 +11,121 @@ import qs from 'qs'
 
 import { ValidationError } from '../../../errors'
 import { Instance } from '../../../instance'
-import { configurableFn, getMediaDuration } from '../../../utilities'
+import { configurable, getMediaDuration } from '../../../utilities'
 import { Request } from '../../requests'
 import { StatusCodes, type IncomingFile } from '../../types'
-import { serverConfigPipe, type Server } from '../base'
-import { base } from '../utils'
+import { Server, serverConfigPipe } from '../base'
 
-export const FastifyServer = configurableFn(serverConfigPipe, (config): Server => {
-	const instance = Instance.get()
-	const app = Fastify({
-		disableRequestLogging: !config.requests.log,
-		loggerInstance: config.requests.log ? instance.log : undefined,
-		ajv: { customOptions: { coerceTypes: false } },
-		routerOptions: {
-			ignoreTrailingSlash: true,
-			caseSensitive: false,
-		},
-		schemaErrorFormatter: (errors, data) =>
-			new ValidationError(
-				errors.map((error) => ({
-					messages: [error.message ?? ''],
-					field: `${data}${error.instancePath}`.replaceAll('/', '.'),
-				})),
-			),
-	})
+export class FastifyServer extends configurable(serverConfigPipe, Server as unknown as new () => Server) {
+	#app: ReturnType<typeof Fastify>
 
-	const server: Server = base(app.server, config, {
-		parseRequest: async (req) => {
-			const { body, files } = excludeBufferKeys(req.body ?? {})
-
-			return new Request({
-				ip: req.ip,
-				body,
-				cookies: req.cookies ?? {},
-				params: req.params ?? <any>{},
-				query: req.query ?? {},
-				method: <any>req.method,
-				path: req.url,
-				headers: req.headers,
-				files,
-			})
-		},
-		handleResponse: async (res, response) => {
-			for (const [key, { value, ...opts }] of Object.entries(response.cookies)) res = res.setCookie(key, value, opts)
-			await res.status(response.status).headers(response.headers).send(response.body)
-		},
-		registerRoute: (method, path, cb) => {
-			app.register(async (inst) => {
-				inst.route({ url: path, method, handler: cb })
-			})
-		},
-		registerErrorHandler: (cb) => {
-			app.setErrorHandler(cb)
-		},
-		registerNotFoundHandler: (cb) => {
-			app.setNotFoundHandler(cb)
-		},
-		start: async (port) => {
-			await app.ready()
-			await app.listen({ port, host: '0.0.0.0' })
-			Instance.on('close', app.close, 1)
-			return true
-		},
-	})
-
-	app.decorateRequest('savedReq', null)
-	app.setValidatorCompiler(() => () => true)
-	app.setSerializerCompiler(() => (data) => JSON.stringify(data))
-	if (config.publicPath) app.register(fastifyStatic, { root: config.publicPath })
-	app.register(fastifyCookie, {})
-	app.register(fastifyCors, server.cors)
-	app.register(fastifyFormBody, { parser: (str) => qs.parse(str) })
-	app.register(fastifyHelmet, { crossOriginResourcePolicy: { policy: 'cross-origin' }, contentSecurityPolicy: false })
-	app.register(fastifyMultipart, {
-		attachFieldsToBody: 'keyValues',
-		throwFileSizeLimit: false,
-		limits: { fileSize: instance.settings.utils.maxFileUploadSizeInMb * 1024 * 1024 },
-		onFile: async (f) => {
-			const buffer = await f.toBuffer()
-			const parsed: IncomingFile = {
-				name: f.filename,
-				type: f.mimetype,
-				size: buffer.byteLength,
-				isTruncated: f.file.truncated,
-				data: buffer,
-				duration: await getMediaDuration(buffer),
-			}
-			// @ts-ignore
-			f.value = parsed
-		},
-	})
-	/* if (this.settings.slowdown.enabled) app.register(fastifySlowDown, {
-			timeWindow: this.settings.slowdown.periodInMs,
-			delayAfter: this.settings.slowdown.delayAfter,
-			delay: this.settings.slowdown.delayInMs
-		}) */
-	if (config.requests.rateLimit.enabled)
-		app.register(fastifyRateLimit, {
-			max: config.requests.rateLimit.limit,
-			timeWindow: config.requests.rateLimit.periodInMs,
-			errorResponseBuilder: (_, context) => ({
-				statusCode: StatusCodes.TooManyRequests,
-				message: JSON.stringify([{ message: `Too Many Requests. Retry in ${context.after}` }]),
-			}),
+	protected constructor(config: typeof FastifyServer.Config) {
+		super(config)
+		const instance = Instance.get()
+		const app = Fastify({
+			disableRequestLogging: !config.requests.log,
+			loggerInstance: config.requests.log ? instance.log : undefined,
+			ajv: { customOptions: { coerceTypes: false } },
+			routerOptions: {
+				ignoreTrailingSlash: true,
+				caseSensitive: false,
+			},
+			schemaErrorFormatter: (errors, data) =>
+				new ValidationError(
+					errors.map((error) => ({
+						messages: [error.message ?? ''],
+						field: `${data}${error.instancePath}`.replaceAll('/', '.'),
+					})),
+				),
 		})
+		this.#app = app
 
-	return server
-})
+		this.setup(app.server, config)
+
+		app.decorateRequest('savedReq', null)
+		app.setValidatorCompiler(() => () => true)
+		app.setSerializerCompiler(() => (data) => JSON.stringify(data))
+		if (config.publicPath) app.register(fastifyStatic, { root: config.publicPath })
+		app.register(fastifyCookie, {})
+		app.register(fastifyCors, this.cors)
+		app.register(fastifyFormBody, { parser: (str) => qs.parse(str) })
+		app.register(fastifyHelmet, { crossOriginResourcePolicy: { policy: 'cross-origin' }, contentSecurityPolicy: false })
+		app.register(fastifyMultipart, {
+			attachFieldsToBody: 'keyValues',
+			throwFileSizeLimit: false,
+			limits: { fileSize: instance.settings.utils.maxFileUploadSizeInMb * 1024 * 1024 },
+			onFile: async (f) => {
+				const buffer = await f.toBuffer()
+				const parsed: IncomingFile = {
+					name: f.filename,
+					type: f.mimetype,
+					size: buffer.byteLength,
+					isTruncated: f.file.truncated,
+					data: buffer,
+					duration: await getMediaDuration(buffer),
+				}
+				// @ts-ignore
+				f.value = parsed
+			},
+		})
+		/* if (this.settings.slowdown.enabled) app.register(fastifySlowDown, {
+				timeWindow: this.settings.slowdown.periodInMs,
+				delayAfter: this.settings.slowdown.delayAfter,
+				delay: this.settings.slowdown.delayInMs
+			}) */
+		if (config.requests.rateLimit.enabled)
+			app.register(fastifyRateLimit, {
+				max: config.requests.rateLimit.limit,
+				timeWindow: config.requests.rateLimit.periodInMs,
+				errorResponseBuilder: (_, context) => ({
+					statusCode: StatusCodes.TooManyRequests,
+					message: JSON.stringify([{ message: `Too Many Requests. Retry in ${context.after}` }]),
+				}),
+			})
+	}
+
+	protected async parseRequest(req: any) {
+		const { body, files } = excludeBufferKeys(req.body ?? {})
+
+		return new Request({
+			ip: req.ip,
+			body,
+			cookies: req.cookies ?? {},
+			params: req.params ?? <any>{},
+			query: req.query ?? {},
+			method: <any>req.method,
+			path: req.url,
+			headers: req.headers,
+			files,
+		})
+	}
+
+	protected async handleResponse(res: any, response: any) {
+		for (const [key, { value, ...opts }] of Object.entries(response.cookies) as [string, any][]) res = res.setCookie(key, value, opts)
+		await res.status(response.status).headers(response.headers).send(response.body)
+	}
+
+	protected registerRoute(method: any, path: string, cb: (req: any, res: any) => Promise<void>) {
+		this.#app.register(async (inst) => {
+			inst.route({ url: path, method, handler: cb })
+		})
+	}
+
+	protected registerErrorHandler(cb: (error: Error, req: any, res: any) => Promise<void>) {
+		this.#app.setErrorHandler(cb)
+	}
+
+	protected registerNotFoundHandler(cb: (req: any, res: any) => Promise<void>) {
+		this.#app.setNotFoundHandler(cb)
+	}
+
+	protected async startServer(port: number) {
+		await this.#app.ready()
+		await this.#app.listen({ port, host: '0.0.0.0' })
+		Instance.on('close', this.#app.close, 1)
+		return true
+	}
+}
 
 function excludeBufferKeys<T>(body: T) {
 	if (typeof body !== 'object') return { body, files: {} }
